the new 'client' common object and the return of the 'create' static method - the EVENT.create method was brought back, can be used instead of EVENT.fire - the common object was seperated into two, now there's a new common object called 'client'

+++ kin-ecosystem-sdk/src/main/java/com/kin/ecosystem/Kin.java
@@ -109,10 +109,10 @@ public class Kin {
         EventsStore.init(new CommonModifier() {
             @Override
             public void modify(CommonProxy mutable) {
-                mutable.setTimestamp(new DynamicValue<Double>() {
+                mutable.setTimestamp(new DynamicValue<Long>() {
                     @Override
-                    public Double get() {
-                        return new Double(System.currentTimeMillis());
+                    public Long get() {
+                        return new Long(System.currentTimeMillis());
                     }
                 });
